show count of users and posts

+++ Backend/controller/categoryController.js
@@ -59,6 +59,10 @@ export const categoryById = async(req,res)=>{
 export const deleteCategory = async (req, res) => {
   try {
     const { id } = req.params;
+    const category_blogs = await Category.findById(id);
+    if(category_blogs.blogs.length>0){
+      return res.status(400).json({ message: "Category has blogs in it. Can't be deleted" });
+    }
     const category = await Category.findByIdAndDelete(id);
     if (!category) {
       return res.status(404).json({ message: "Category not found" });
